Function wayland_window::init_decorated_surface

pub fn init_decorated_surface<ID: 'static>(
    evqh: &mut EventQueueHandle, 
    implementation: DecoratedSurfaceImplementation<ID>, 
    idata: ID, 
    surface: &WlSurface, 
    width: i32, 
    height: i32, 
    compositor: &WlCompositor, 
    subcompositor: &WlSubcompositor, 
    shm: &WlShm, 
    shell: &Shell, 
    seat: Option<WlSeat>, 
    decorate: bool
) -> Result<DecoratedSurface, ()>

Creates a new DecoratedSurface and inserts it in the event queue

This requires you to provide:

a mutable reference to the event_queue

an implementation (and implementation data) for the decorated surface

a reference to the surface to decorate

the wanted dimensions for this decoration

a few wayland globals: compositor, subcompositor, shm, shell, and an optional seat

a bool specifying if decorations should be displayed or not

If you provide a seat, it will be used to process user interaction with the decorations.