Struct wayland_window::FrameImplementation

pub struct FrameImplementation<ID> {
    pub configure: fn(evqh: &mut EventQueueHandle, idata: &mut ID, cfg: Configure, newsize: Option<(i32, i32)>),
    pub close: fn(evqh: &mut EventQueueHandle, idata: &mut ID),
    pub refresh: fn(evqh: &mut EventQueueHandle, idata: &mut ID),
}

For handling events that occur to a Frame.

Fields
configure: fn(evqh: &mut EventQueueHandle, idata: &mut ID, cfg: Configure, newsize: Option<(i32, i32)>)

Called whenever the Frame has been resized.

Note: if you've not set a minimum size, width and height will not always be positive values. Values can be negative if a user attempts to resize the window past the left or top borders.

close: fn(evqh: &mut EventQueueHandle, idata: &mut ID)

Called when the Frame is closed.

refresh: fn(evqh: &mut EventQueueHandle, idata: &mut ID)

Called when the Frame wants to be refreshed