Function wayland_window::create_frame

pub fn create_frame<ID: 'static>(
    evqh: &mut EventQueueHandle, 
    implementation: FrameImplementation<ID>, 
    idata: ID, 
    surface: &WlSurface, 
    width: i32, 
    height: i32, 
    compositor: &WlCompositor, 
    subcompositor: &WlSubcompositor, 
    shm: &WlShm, 
    shell: &Shell, 
    seat: Option<WlSeat>
) -> Result<Frame, ()>

Create a decoration frame for a wl_surface

This will create a decoration and declare it as a shell surface to the wayland compositor.

See crate documentations for details about how to use it.