Struct wayland_window::DecoratedSurface

pub struct DecoratedSurface<S: Surface> {
    // some fields omitted
}

A decorated surface, wrapping a wayalnd surface and handling its decorations.

Methods

impl<S: Surface> DecoratedSurface<S>

fn new(user_surface: S, registry: &Registry) -> Result<DecoratedSurface<S>, S>

Creates a new decorated window around given surface.

If the creation failed (likely if the registry was not ready), hands back the surface.

fn unwrap(self) -> S

Destroys the decorated window and gives back the wrapped surface.